+++ crates/biome_js_analyze/Cargo.toml
@@ -24,6 +24,7 @@ biome_js_syntax          = { workspace = true }
 biome_project            = { workspace = true }
 biome_rowan              = { workspace = true }
 biome_string_case        = { workspace = true }
+biome_suppression        = { workspace = true }
 biome_unicode_table      = { workspace = true }
 lazy_static              = { workspace = true }
 natord                   = "1.0.9"
